add example data

+++ routes/upload.js
@@ -34,23 +34,35 @@ exports.upload = function(req, res) {
     form.encoding = 'utf-8';
     form.uploadDir = "uploads/videos/";
     form.parse(req, function(err, fields, files){
-        var uploads = files.video;
-        console.log(files);
-        cmdStr += ' ' + uploads[0].originalFilename + ' 0';
-        fs.renameSync(uploads[0].path, form.uploadDir + uploads[0].originalFilename);
-        exec(cmdStr);
-        checkFinish(function () {
-            var plotData = fs.readFileSync(generatePath('plot_data.txt'), 'utf-8');
-            var sliceData = fs.readFileSync(generatePath('data_slice.txt'), 'utf-8');
-            var cmdClear = 'rm ' + generatePath('finish.txt');
-            exec(cmdClear);
+        if (fields.angle == -1) {
+            var plotData = fs.readFileSync('uploads/plot_data.txt', 'utf-8');
+            var sliceData = fs.readFileSync('uploads/data_slice.txt', 'utf-8');
             res.status(200).send({
                 success: true,
                 data: {
-                  total: plotData,
-                  slice: sliceData
+                    total: plotData,
+                    slice: sliceData
                 }
             });
-        });
+        } else {
+            var uploads = files.video;
+            cmdStr += ' ' + uploads[0].originalFilename + ' 0';
+            fs.renameSync(uploads[0].path, form.uploadDir + uploads[0].originalFilename);
+
+            exec(cmdStr);
+            checkFinish(function () {
+                var plotData = fs.readFileSync(generatePath('plot_data.txt'), 'utf-8');
+                var sliceData = fs.readFileSync(generatePath('data_slice.txt'), 'utf-8');
+                var cmdClear = 'rm ' + generatePath('finish.txt');
+                exec(cmdClear);
+                res.status(200).send({
+                    success: true,
+                    data: {
+                        total: plotData,
+                        slice: sliceData
+                    }
+                });
+            });
+        }
     });
 };
